MSVC: Avoid warning C4267 (model side)
Add explicit casts from size_type (std::vector::size()) to int, as this
would otherwise trigger the C4267 warning from MSVC compiler about
possible loss of data.

It is safe to assume that the size of these vectors is always less
than INT_MAX.

Note: There are still more occurrences of this warning in the kernel
      sources.  However, this commit avoids the need to suppress the
      warning on the model side.

+++ src/sysc/communication/sc_export.h
@@ -223,7 +223,7 @@ public:
     void remove( sc_export_base* );
 
     int size() const
-        { return m_export_vec.size(); }
+        { return static_cast<int>(m_export_vec.size()); }
 
 private:
